# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CPA Planning Department Interactive Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  employee: 'employee',
  officer: 'officer',
  manager: 'manager',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  role: UserRole;
  active: boolean;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  employee: 'employee',
  officer: 'officer',
  manager: 'manager',
  admin: 'admin',
} as const;

export interface UserInput {
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  employee: 'employee',
  officer: 'officer',
  manager: 'manager',
  admin: 'admin',
} as const;

export interface UserUpdate {
  nameAr?: string;
  /** @nullable */
  nameEn?: string | null;
  role?: UserUpdateRole;
  active?: boolean;
}

export type AnnouncementCategory = typeof AnnouncementCategory[keyof typeof AnnouncementCategory];


export const AnnouncementCategory = {
  announcement: 'announcement',
  update: 'update',
  circular: 'circular',
  deadline: 'deadline',
} as const;

export interface Announcement {
  id: number;
  title: string;
  body: string;
  category: AnnouncementCategory;
  date: string;
  authorId: number;
  authorName: string;
  archived: boolean;
}

export type AnnouncementInputCategory = typeof AnnouncementInputCategory[keyof typeof AnnouncementInputCategory];


export const AnnouncementInputCategory = {
  announcement: 'announcement',
  update: 'update',
  circular: 'circular',
  deadline: 'deadline',
} as const;

export interface AnnouncementInput {
  title: string;
  body: string;
  category: AnnouncementInputCategory;
  authorId: number;
}

export type AnnouncementUpdateCategory = typeof AnnouncementUpdateCategory[keyof typeof AnnouncementUpdateCategory];


export const AnnouncementUpdateCategory = {
  announcement: 'announcement',
  update: 'update',
  circular: 'circular',
  deadline: 'deadline',
} as const;

export interface AnnouncementUpdate {
  title?: string;
  body?: string;
  category?: AnnouncementUpdateCategory;
  archived?: boolean;
}

export type DiscussionStatus = typeof DiscussionStatus[keyof typeof DiscussionStatus];


export const DiscussionStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface Discussion {
  id: number;
  title: string;
  description: string;
  status: DiscussionStatus;
  date: string;
  authorId: number;
  authorName: string;
  commentCount: number;
}

export type DiscussionDetailStatus = typeof DiscussionDetailStatus[keyof typeof DiscussionDetailStatus];


export const DiscussionDetailStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface Comment {
  id: number;
  discussionId: number;
  userId: number;
  userName: string;
  isStaff: boolean;
  text: string;
  date: string;
}

export interface DiscussionDetail {
  id: number;
  title: string;
  description: string;
  status: DiscussionDetailStatus;
  date: string;
  authorId: number;
  authorName: string;
  comments: Comment[];
}

export interface DiscussionInput {
  title: string;
  description: string;
  authorId: number;
}

export type DiscussionUpdateStatus = typeof DiscussionUpdateStatus[keyof typeof DiscussionUpdateStatus];


export const DiscussionUpdateStatus = {
  open: 'open',
  closed: 'closed',
} as const;

export interface DiscussionUpdate {
  status?: DiscussionUpdateStatus;
  title?: string;
  description?: string;
}

export interface CommentInput {
  userId: number;
  text: string;
}

export type InquiryStatus = typeof InquiryStatus[keyof typeof InquiryStatus];


export const InquiryStatus = {
  open: 'open',
  answered: 'answered',
  resolved: 'resolved',
} as const;

export interface Inquiry {
  id: number;
  userId: number;
  userName: string;
  subject: string;
  details: string;
  status: InquiryStatus;
  date: string;
  /** @nullable */
  response?: string | null;
  /** @nullable */
  responderId?: number | null;
  /** @nullable */
  responderName?: string | null;
}

export interface InquiryInput {
  userId: number;
  subject: string;
  details: string;
}

export type InquiryUpdateStatus = typeof InquiryUpdateStatus[keyof typeof InquiryUpdateStatus];


export const InquiryUpdateStatus = {
  open: 'open',
  answered: 'answered',
  resolved: 'resolved',
} as const;

export interface InquiryUpdate {
  status?: InquiryUpdateStatus;
  response?: string;
  responderId?: number;
}

export type DocumentCategory = typeof DocumentCategory[keyof typeof DocumentCategory];


export const DocumentCategory = {
  manuals: 'manuals',
  guidelines: 'guidelines',
  kpi: 'kpi',
  policies: 'policies',
  templates: 'templates',
  other: 'other',
} as const;

export interface Document {
  id: number;
  name: string;
  category: DocumentCategory;
  description: string;
  date: string;
  /** @nullable */
  fileUrl?: string | null;
}

export type DocumentInputCategory = typeof DocumentInputCategory[keyof typeof DocumentInputCategory];


export const DocumentInputCategory = {
  manuals: 'manuals',
  guidelines: 'guidelines',
  kpi: 'kpi',
  policies: 'policies',
  templates: 'templates',
  other: 'other',
} as const;

export interface DocumentInput {
  name: string;
  category: DocumentInputCategory;
  description: string;
  /** @nullable */
  fileUrl?: string | null;
}

export interface Faq {
  id: number;
  question: string;
  answer: string;
}

export interface FaqInput {
  question: string;
  answer: string;
}

export interface FaqUpdate {
  question?: string;
  answer?: string;
}

export type SuggestionCategory = typeof SuggestionCategory[keyof typeof SuggestionCategory];


export const SuggestionCategory = {
  improvement: 'improvement',
  initiative: 'initiative',
  process: 'process',
  feedback: 'feedback',
} as const;

export type SuggestionStatus = typeof SuggestionStatus[keyof typeof SuggestionStatus];


export const SuggestionStatus = {
  new: 'new',
  under_review: 'under_review',
  accepted: 'accepted',
  rejected: 'rejected',
  implemented: 'implemented',
} as const;

export interface Suggestion {
  id: number;
  userId: number;
  userName: string;
  category: SuggestionCategory;
  text: string;
  status: SuggestionStatus;
  date: string;
  /** @nullable */
  feedback?: string | null;
}

export type SuggestionInputCategory = typeof SuggestionInputCategory[keyof typeof SuggestionInputCategory];


export const SuggestionInputCategory = {
  improvement: 'improvement',
  initiative: 'initiative',
  process: 'process',
  feedback: 'feedback',
} as const;

export interface SuggestionInput {
  userId: number;
  category: SuggestionInputCategory;
  text: string;
}

export type SuggestionUpdateStatus = typeof SuggestionUpdateStatus[keyof typeof SuggestionUpdateStatus];


export const SuggestionUpdateStatus = {
  new: 'new',
  under_review: 'under_review',
  accepted: 'accepted',
  rejected: 'rejected',
  implemented: 'implemented',
} as const;

export interface SuggestionUpdate {
  status?: SuggestionUpdateStatus;
  feedback?: string;
}

export interface DashboardStats {
  openInquiries: number;
  newSuggestions: number;
  activeDiscussions: number;
  totalDocuments: number;
  recentAnnouncements: number;
  totalUsers: number;
}

export type ListAnnouncementsParams = {
archived?: boolean;
};

export type ListInquiriesParams = {
userId?: number;
};

export type ListDocumentsParams = {
category?: string;
};

export type ListSuggestionsParams = {
userId?: number;
};