# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CPA Planning Department Interactive Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: archived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement (archive/restore)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /discussions:
    get:
      operationId: listDiscussions
      tags: [discussions]
      summary: List discussions
      responses:
        "200":
          description: List of discussions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discussion"
    post:
      operationId: createDiscussion
      tags: [discussions]
      summary: Create a discussion topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
  /discussions/{id}:
    get:
      operationId: getDiscussion
      tags: [discussions]
      summary: Get a discussion with its comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussion detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussionDetail"
    patch:
      operationId: updateDiscussion
      tags: [discussions]
      summary: Update discussion status (open/closed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
  /discussions/{id}/comments:
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment on a discussion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List knowledge base documents
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add a document to the knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Announcement:
      type: object
      required: [id, title, body, category, date, authorId, authorName, archived]
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        category:
          type: string
          enum: [announcement, update, circular, deadline]
        date:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        archived:
          type: boolean
    AnnouncementInput:
      type: object
      required: [title, body, category, authorId]
      properties:
        title:
          type: string
        body:
          type: string
        category:
          type: string
          enum: [announcement, update, circular, deadline]
        authorId:
          type: integer
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        category:
          type: string
          enum: [announcement, update, circular, deadline]
        archived:
          type: boolean
    Discussion:
      type: object
      required: [id, title, description, status, date, authorId, authorName, commentCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, closed]
        date:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        commentCount:
          type: integer
    DiscussionDetail:
      type: object
      required: [id, title, description, status, date, authorId, authorName, comments]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, closed]
        date:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    DiscussionInput:
      type: object
      required: [title, description, authorId]
      properties:
        title:
          type: string
        description:
          type: string
        authorId:
          type: integer
    DiscussionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, closed]
        title:
          type: string
        description:
          type: string
    Comment:
      type: object
      required: [id, discussionId, userId, userName, isStaff, text, date]
      properties:
        id:
          type: integer
        discussionId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        isStaff:
          type: boolean
        text:
          type: string
        date:
          type: string
    CommentInput:
      type: object
      required: [userId, text]
      properties:
        userId:
          type: integer
        text:
          type: string
...
    Document:
      type: object
      required: [id, name, category, description, date]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [manuals, guidelines, kpi, policies, templates, other]
        description:
          type: string
        date:
          type: string
        fileUrl:
          type: ["string", "null"]
    DocumentInput:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [manuals, guidelines, kpi, policies, templates, other]
        description:
          type: string
        fileUrl:
          type: ["string", "null"]
...