# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CPA Planning Department Interactive Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementUpdateCategory } from './announcementUpdateCategory';

export interface AnnouncementUpdate {
  title?: string;
  body?: string;
  category?: AnnouncementUpdateCategory;
  archived?: boolean;
}