# lib/api-zod/src/generated/types/discussionDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CPA Planning Department Interactive Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from './comment';
import type { DiscussionDetailStatus } from './discussionDetailStatus';

export interface DiscussionDetail {
  id: number;
  title: string;
  description: string;
  status: DiscussionDetailStatus;
  date: string;
  authorId: number;
  authorName: string;
  comments: Comment[];
}