# lib/api-zod/src/generated/types/discussionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CPA Planning Department Interactive Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscussionInput {
  title: string;
  description: string;
  authorId: number;
}